refactor(layout): extract some shared layout styling info
Useful prerequisite to make the upcoming PR for the user menu smaller
