+++ packages/react-components/src/__tests__/pixels.test.ts
@@ -1,12 +1,4 @@
-import {
-  screen,
-  vminLinearCalc,
-  Screen,
-  contentSidePaddingWithNavigation,
-  contentSidePaddingWithoutNavigation,
-  largeDesktopScreen,
-} from '../pixels';
-import { viewportCalc } from '../test-utils';
+import { screen, vminLinearCalc, Screen } from '../pixels';
 
 describe('screen', () => {
   it('sets the dimensions', () => {
@@ -26,53 +18,5 @@ describe('vminLinearCalc', () => {
     expect(vminLinearCalc(smallScreen, 10, largeScreen, 12, 'px')).toEqual(
       'calc(8px + 2vmin)',
     );
-  });
-});
-
-describe('contentSidePaddingWithoutNavigation', () => {
-  it('without arguments returns the minimum side padding', () => {
-    expect(
-      viewportCalc(contentSidePaddingWithoutNavigation(), largeDesktopScreen),
-    ).toMatchInlineSnapshot(`"159px"`);
-  });
-
-  it('increases the padding to account for fewer columns', () => {
-    const normalSidePadding = Number(
-      viewportCalc(
-        contentSidePaddingWithoutNavigation(),
-        largeDesktopScreen,
-      ).replace(/px$/, ''),
-    );
-    const twoColumnSidePadding = Number(
-      viewportCalc(
-        contentSidePaddingWithoutNavigation(2),
-        largeDesktopScreen,
-      ).replace(/px$/, ''),
-    );
-    expect(twoColumnSidePadding).toBeGreaterThan(normalSidePadding);
-  });
-});
-
-describe('contentSidePaddingWithNavigation', () => {
-  it('without arguments returns the minimum side padding', () => {
-    expect(
-      viewportCalc(contentSidePaddingWithNavigation(), largeDesktopScreen),
-    ).toMatchInlineSnapshot(`"30px"`);
-  });
-
-  it('increases the padding to account for fewer columns', () => {
-    const normalSidePadding = Number(
-      viewportCalc(
-        contentSidePaddingWithNavigation(),
-        largeDesktopScreen,
-      ).replace(/px$/, ''),
-    );
-    const twoColumnSidePadding = Number(
-      viewportCalc(
-        contentSidePaddingWithNavigation(2),
-        largeDesktopScreen,
-      ).replace(/px$/, ''),
-    );
-    expect(twoColumnSidePadding).toBeGreaterThan(normalSidePadding);
   });
 });
