+++ packages/react-components/src/pixels.ts
@@ -36,32 +36,3 @@ export const vminLinearCalc = (
 };
 
 export const formTargetWidth = 354;
-
-export const layoutDrawerQuery = `@media (max-width: ${
-  smallDesktopScreen.width - 1
-}px)`;
-export const layoutCrossQuery = `@media (min-width: ${smallDesktopScreen.width}px)`;
-
-const largeDesktopColWidth = 66;
-const largeDesktopColGap = 30;
-export const contentSidePaddingWithoutNavigation = (
-  desktopCols: 2 | 4 | 6 | 8 | 10 | 12 = 12,
-): string =>
-  vminLinearCalc(
-    mobileScreen,
-    24,
-    largeDesktopScreen,
-    159 +
-      ((12 - desktopCols) / 2) * (largeDesktopColWidth + largeDesktopColGap),
-    'px',
-  );
-export const contentSidePaddingWithNavigation = (
-  desktopCols: 2 | 4 | 6 | 8 | 10 | 12 = 12,
-): string =>
-  vminLinearCalc(
-    mobileScreen,
-    24,
-    largeDesktopScreen,
-    30 + ((12 - desktopCols) / 2) * (largeDesktopColWidth + largeDesktopColGap),
-    'px',
-  );
