+++ packages/react-components/src/templates/TeamAbout.tsx
@@ -1,9 +1,10 @@
 import React, { ComponentProps } from 'react';
 import css from '@emotion/css';
 
-import { perRem, contentSidePaddingWithNavigation } from '../pixels';
+import { perRem } from '../pixels';
 import { pearl, steel } from '../colors';
 import { MembersSection, SkillsSection, TeamOverview } from '../organisms';
+import { contentSidePaddingWithNavigation } from '../layout';
 
 const styles = css({
   backgroundColor: pearl.rgb,
